refactor: split config types module
